Homepage and Events Adjustments (#46)
* fix(homepage layout): fit animation in flexbox

Added css variables and implemented useRef to dynamically set the
height for lottie animation and homepage containers based on
viewport height.

* fix(events page): fit to single page

Adjust typograpghy to fit on a single page since Poppins font-family is
larger than default. Tried to keep with the sizing and weight of
original page.

* fix(layouts): home and events pages

Remove the logic from home.tsx. Revert to using CSS instead. Adjust
layout so navbar and footer are always visible. Shrink events
description to max-width of 600px for UX on wide screens.