Font optimization (#48)
* fix(homepage layout): fit animation in flexbox

Added css variables and implemented useRef to dynamically set the
height for lottie animation and homepage containers based on
viewport height.

* fix(events page): fit to single page

Adjust typograpghy to fit on a single page since Poppins font-family is
larger than default. Tried to keep with the sizing and weight of
original page.

* fix(layouts): home and events pages

Remove the logic from home.tsx. Revert to using CSS instead. Adjust
layout so navbar and footer are always visible. Shrink events
description to max-width of 600px for UX on wide screens.

* build(yarn): update to latest version of Next

In order to use next/font, Next had to be updated.
While allowing the fonts to be served correctly, there are other
warnings server side that will need to be addressed regarding
images.domains and react-I18Next.

* perf(next.config): Implement remotePatterns

Replace domains property with remotePatterns for external images from
github. Server warned that domains was deprecated and remotePatterns is
the new property to use.

* refactor(fonts): revert back to Poppins

Place font injection in Layout component instead of a fragment wrapper
in _app.tsx.