Members (#50)
* fix(members): add heading tag

Change p to h1 since this is the title of the page.

* feat(members): center content

Center the title and the member cards given space between.

* feat(members): center member cards

Actually, center had better feel since all cards with have the same gap
between them.

* feat(page headings): consistent font-size

Consistent font-sizes for the first element on each page.

* fix(members): dynamic member list

There is a Google Sheets API for the member data. The page fetches the
data. It needs to be optimized, though.

* perf(members): attempting to load members

Loading is slow so it needs an indicator for UX

* fix(members): member data loading immediately

Use getStaticProps with revalidate set to one week. We can see how it
work and then update accordingly.

* feat(events): consistent font sizes

Make font-size in media-query for events the same as other pages

* chore(loading): remove unused loading component

* fix(members): type assertion and revalidate time

remove type assertion and shorten revalidate time.